Add test codes
